feat(server): ability to extract data deep from the response
