+++ src/ng2-smart-table/lib/data-source/server/server.data-source.ts
@@ -4,6 +4,7 @@ import { RequestOptionsArgs } from '@angular/http/src/interfaces';
 import { URLSearchParams } from '@angular/http/src/url_search_params';
 import { Observable } from 'rxjs';
 import { ServerSourceConf } from './server-source.conf';
+import { getDeepFromObject } from '../../helpers';
 
 export class ServerDataSource extends LocalDataSource {
 
@@ -34,12 +35,35 @@ export class ServerDataSource extends LocalDataSource {
     }).toPromise();
   }
 
+  /**
+   * Extracts array of data from server response
+   * @param res
+   * @returns {any}
+   */
   protected extractDataFromResponse(res): Array<any> {
-    return res.json();
+    const rawData = res.json();
+    let data = !!this.conf.dataKey ? getDeepFromObject(rawData, this.conf.dataKey, []) : rawData;
+
+    if (data instanceof Array) {
+      return data;
+    }
+
+    throw new Error(`Data must be an array. Please check that data extracted from the server response by the key '${this.conf.dataKey}' exists and is array.`);
   }
 
+  /**
+   * Extracts total rows count from the server response
+   * Looks for the count in the heders first, then in the response body
+   * @param res
+   * @returns {any}
+   */
   protected extractTotalFromResponse(res): number {
-    return +res.headers.get(this.conf.totalKey);
+    if (res.headers.has(this.conf.totalKey)) {
+      return +res.headers.get(this.conf.totalKey);
+    } else {
+      const rawData = res.json();
+      return getDeepFromObject(rawData, this.conf.totalKey, [])
+    }
   }
 
   protected requestElements(): Observable<any> {
@@ -71,9 +95,12 @@ export class ServerDataSource extends LocalDataSource {
   protected addFilterRequestOptions(requestOptions: RequestOptionsArgs): RequestOptionsArgs {
     let searchParams: URLSearchParams = <URLSearchParams> requestOptions.search;
 
-    if (this.pagingConf && this.pagingConf['page'] && this.pagingConf['perPage']) {
-      searchParams.set(this.conf.pagerPageKey, this.pagingConf['page']);
-      searchParams.set(this.conf.pagerLimitKey, this.pagingConf['perPage']);
+    if (this.filterConf.filters) {
+      this.filterConf.filters.forEach((fieldConf) => {
+        if (fieldConf['search']) {
+          searchParams.set(this.conf.filterFieldKey.replace('#field#', fieldConf['field']), fieldConf['search']);
+        }
+      });
     }
 
     return requestOptions;
@@ -82,12 +109,9 @@ export class ServerDataSource extends LocalDataSource {
   protected addPagerRequestOptions(requestOptions: RequestOptionsArgs): RequestOptionsArgs {
     let searchParams: URLSearchParams = <URLSearchParams> requestOptions.search;
 
-    if (this.filterConf.filters) {
-      this.filterConf.filters.forEach((fieldConf) => {
-        if (fieldConf['search']) {
-          searchParams.set(this.conf.filterFieldKey.replace('#field#', fieldConf['field']), fieldConf['search']);
-        }
-      });
+    if (this.pagingConf && this.pagingConf['page'] && this.pagingConf['perPage']) {
+      searchParams.set(this.conf.pagerPageKey, this.pagingConf['page']);
+      searchParams.set(this.conf.pagerLimitKey, this.pagingConf['perPage']);
     }
 
     return requestOptions;
